Fixed issue #751 and using Xcode 3.2.3 as default
Issue #751: fixed artifacts in PageTurnTransition
Templates: Using Xcode 3.2.3 + SDK 4.0 + Deplay OS Target 3.0 as default

Former-commit-id: 6e57ee3d523f15f068786b79cffe6e43461478c9

+++ tests/bugs/Bug-422.m
@@ -57,7 +57,7 @@
 	
 	CCArray *array = [t children];
     for (CCNode *node in array) {
-        NSLog(@"0x%x, rc: %d", node, [node retainCount]);
+        NSLog(@"0x%x, rc: %d", (unsigned int)node, [node retainCount]);
         [self check:node];
     }
 }
